pyproject.toml
[tool.poetry]
name = "figure_hook"
version = "0.0.5"
description = ""
authors = ["Elton Chou <plscd748@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
figure-parser = {git = "https://github.com/FigureHook/figure_parser.git", tag = "v0.1.0"}
"discord.py" = "^1.7.3"
plurk-oauth = "^0.9.2"
requests = "^2.26.0"
sqlalchemy-mixins = "^1.5"
Babel = "^2.9.1"
psycopg2-binary = "^2.9.1"
scrapyd-client = "^1.2.0"
cryptography = "^36.0.0"
SQLAlchemy = {extras = ["mypy"], version = "^1.4.32"}

[tool.poetry.dev-dependencies]
coverage = "^5.5"
Faker = "^8.13.2"
pytest = "^6.2.5"
pytest-mock = "^3.6.1"
flake8 = "^3.9.2"
autopep8 = "^1.5.7"
alembic = "^1.7.1"
types-pytz = "^2021.3.6"
types-requests = "^2.27.14"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
